% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.c50-package}
\alias{mlr3learners.c50}
\alias{mlr3learners.c50-package}
\title{mlr3learners.c50: C5.0 Learner for 'mlr3'}
\description{
C5.0 Learner for 'mlr3'.
}
\author{
\strong{Maintainer}: Henri Funk \email{henrifnk@gmail.com}

}
